part2/worker/requirements.txt
redis==5.0.0
psycopg2-binary==2.9.9
